Intermediate results: succesful subgraph extraction multiplication and insertion
