+++ regelum/environment/node/base.py
@@ -7,7 +7,6 @@ The module includes:
     - State: Wrapper for hierarchical state management
     - Inputs: Handler for node input dependencies
     - Node: Base class for all computational nodes
-    - Graph: Manages node execution and dependencies
     - Clock: Provides timing functionality
     - Logger: Records state evolution
     - MPCNode: Implements Model Predictive Control
@@ -54,157 +53,6 @@ from regelum.environment.transistor import (
     SampleAndHoldModifier,
 )
 from typing import Optional, List
-from graphviz import Digraph
-
-
-def visualize_graph(
-    graph: "Graph", output_file: Optional[str] = "graph_diagram", view: bool = True
-):
-    """Visualize the given Graph instance as a block diagram using graphviz.
-
-    Each Node is represented as a box with:
-      - "ClassName:root_state_name" as title (bold)
-      - Node type (Continuous or Discrete)
-      - Initial state values (if leaf and defined)
-      - A table showing the transistor and its modifiers (if any).
-        Modifiers are detected by a class attribute `is_modifier = True`.
-        The main transistor class is one of ODETransistor, ScipyTransistor, CasADiTransistor,
-        or Transistor if none of these specialized classes appear.
-
-    Edges represent dependencies between nodes based on their input states and are labeled:
-      - The labels are bold, slightly larger, with a white background.
-    """
-    dot = Digraph(comment="Regelum Graph Visualization", format="pdf")
-    dot.attr(rankdir="LR", concentrate="true", nodesep="0.5", ranksep="0.7")
-
-    nodes = graph.nodes
-    node_ids = {node: f"node_{i}" for i, node in enumerate(nodes)}
-
-    def prettify_value(val):
-        if isinstance(val, np.ndarray):
-            return f"{val}"
-        elif hasattr(val, "__class__"):
-            return val.__class__.__name__
-        else:
-            return str(val)
-
-    def get_transistor_info(transistor):
-        if not transistor:
-            return None, [], False
-
-        mro = type(transistor).mro()
-        modifiers = []
-        main_transistor_cls = None
-        has_reset = False
-
-        # First pass - collect all modifiers and check for reset
-        for cls in mro:
-            if cls is object:
-                continue
-            if getattr(cls, "is_modifier", False):
-                if cls.__name__ == "ResetTransistor":
-                    has_reset = True
-                else:
-                    modifiers.append(cls.__name__)
-
-        # Second pass - find main transistor class
-        for cls in mro:
-            if cls is object or cls.__name__ == "Transistor":
-                continue
-            if not getattr(cls, "is_modifier", False):  # Only consider non-modifiers
-                main_transistor_cls = cls.__name__
-                break
-
-        if main_transistor_cls is None:
-            main_transistor_cls = "Transistor"
-
-        modifiers.reverse()
-        return main_transistor_cls, modifiers if modifiers else None, has_reset
-
-    # Render each node
-    for node in nodes:
-        node_id = node_ids[node]
-        state_name = node.state.name
-        node_class_name = node.__class__.__name__
-        node_type_info = (
-            "Transition: Continuous" if node.is_continuous else "Transition: Discrete"
-        )
-
-        init_val_str = ""
-        if node.state.is_leaf and node.state.data is not None:
-            val_str = prettify_value(node.state.data)
-            init_val_str = f"Init state: {val_str}"
-
-        main_transistor_cls, modifiers, has_reset = get_transistor_info(node.transistor)
-
-        # Start main table
-        label = f"""<<table border='0' cellborder='0' cellspacing='0'>"""
-
-        label += f"""<tr><td><b>{node_class_name}:{state_name}</b></td></tr>
-            <tr><td>{node_type_info}</td></tr>"""
-
-        if init_val_str:
-            label += f"<tr><td>{init_val_str}</td></tr>"
-
-        if main_transistor_cls:
-            label += "<tr><td><table border='1' cellborder='1' cellspacing='0'>"
-            if modifiers:
-                for modifier in modifiers:
-                    label += f"<tr><td bgcolor='#d9d9d9'>{modifier}</td></tr>"
-            label += f"<tr><td bgcolor='#e6e6e6'>{main_transistor_cls}</td></tr>"
-            label += "</table></td></tr>"
-
-        # Add arrow-like reset indicator at the bottom if needed
-        if has_reset:
-            label += """<tr><td align="center">
-                <table border='0' cellborder='0' cellpadding='2'>
-                    <tr>
-                        <td bgcolor="#ffcccc" border='1' sides="lt" style="rounded"><font point-size="12"><b>➜ Reset</b></font></td>
-                    </tr>
-                </table>
-            </td></tr>"""
-
-        label += "</table>>"
-
-        dot.node(
-            node_id,
-            label=label,
-            shape="box",
-            style="filled",
-            fillcolor="#f0f0f0",
-            margin="0.3",
-        )
-
-    # Map full_path -> owner node
-    state_to_owner = {}
-    for n in nodes:
-        for s in n.state.get_all_states():
-            if s.is_leaf:
-                full_path = s.paths[0]
-                state_to_owner[full_path] = n
-
-    # Add edges
-    for node in nodes:
-        for input_state in node.inputs.states:
-            if input_state.is_leaf:
-                input_path = input_state.paths[0]
-                producer = state_to_owner.get(input_path)
-                if producer and producer != node:
-                    edge_label = f"""<<table border='0' cellborder='0'>
-                        <tr><td bgcolor='white'><font point-size='12'><b>{input_state.name}</b></font></td></tr>
-                        </table>>"""
-
-                    dot.edge(
-                        node_ids[producer],
-                        node_ids[node],
-                        label=edge_label,
-                        fontsize="10",
-                        labeldistance="0.2",
-                        labelangle="0",
-                        labelfloat="false",
-                    )
-
-    dot.render(output_file, view=view)
 
 
 T = TypeVar("T")
@@ -580,302 +428,6 @@ class Node(ABC):
                 found_states.add(path)
 
         return found_states
-
-
-class Graph:
-    """Manages node execution order and dependencies.
-
-    Args:
-        nodes: List of nodes to manage
-        states_to_log: State paths to record
-        logger_cooldown: Minimum time between logs
-    """
-
-    def __init__(
-        self,
-        nodes: List[Node],
-        states_to_log: Optional[List[str]] = None,
-        logger_cooldown: float = 0.0,
-    ) -> None:
-        """Initialize the Graph.
-
-        Args:
-            nodes: List of nodes to manage.
-            states_to_log: State paths to record.
-            logger_cooldown: Minimum time between logs.
-        """
-        fundamental_step_size = self._validate_and_set_step_sizes(nodes)
-        self._setup_logger(nodes, states_to_log, logger_cooldown, fundamental_step_size)
-        self._initialize_graph(nodes, fundamental_step_size)
-
-    def define_fundamental_step_size(self, nodes: List[Node]):
-        step_sizes = [
-            node.step_size
-            for node in nodes
-            if not node.is_continuous and node.step_size is not None
-        ]
-
-        def float_gcd(a: float, b: float) -> float:
-            precision = 1e-9
-            a, b = round(a / precision), round(b / precision)
-            return gcd(int(a), int(b)) * precision
-
-        fundamental_step_size = (
-            reduce(float_gcd, step_sizes) if len(set(step_sizes)) > 1 else step_sizes[0]
-        )
-        return fundamental_step_size
-
-    def _validate_and_set_step_sizes(self, nodes: List[Node]):
-        defined_step_sizes = [
-            node.step_size for node in nodes if node.step_size is not None
-        ]
-        if not defined_step_sizes:
-            raise ValueError("At least one node must have a defined step_size")
-
-        fundamental_step_size = self.define_fundamental_step_size(nodes)
-        for node in nodes:
-            if node.step_size is None:
-                node.step_size = fundamental_step_size
-
-        return fundamental_step_size
-
-    def _setup_logger(
-        self,
-        nodes: List[Node],
-        states_to_log: Optional[List[str]],
-        logger_cooldown: float,
-        fundamental_step_size: float,
-    ):
-        if not states_to_log:
-            self.logger = None
-            return
-
-        self.logger = Logger(
-            states_to_log, fundamental_step_size, cooldown=logger_cooldown
-        )
-        nodes.append(self.logger)
-
-    def _get_logger_step_sizes(
-        self, nodes: List[Node], states_to_log: List[str]
-    ) -> List[float]:
-        return [
-            node.step_size
-            for node in nodes
-            if node.step_size is not None
-            and any(
-                state.search_by_path(path)
-                for state in node.state.get_all_states()
-                for path in states_to_log
-            )
-        ]
-
-    def _initialize_graph(self, nodes: List[Node], fundamental_step_size: float):
-        self.nodes = nodes + [Clock(fundamental_step_size), StepCounter(nodes)]
-        states = reduce(
-            lambda x, y: x + y, [node.state.get_all_states() for node in self.nodes]
-        )
-
-        # Find nodes that need reset modification
-        reset_map = {}
-        for node in self.nodes:
-            for input_path in node.state.paths:
-                if input_path.startswith("reset_"):
-
-                    target_node_name = input_path[6:]  # Remove 'reset_' prefix
-                    reset_map[target_node_name] = True
-
-        from regelum.environment.transistor import ResetModifier
-
-        reset_modifier = ResetModifier()
-
-        # Apply reset modifiers where needed
-        for node in self.nodes:
-            if node.state.name in reset_map:
-                if node.transistor is None:
-                    node.default_transistor_configuration["transistor"] = (
-                        reset_modifier.apply_class(
-                            node.default_transistor_configuration["transistor"]
-                        )
-                    )
-        for node in self.nodes:
-            node.inputs.resolve(states)
-
-        self.ordered_nodes = self.resolve(self.nodes)
-        self._log_node_order()
-
-        for node in self.ordered_nodes:
-            if node.transistor is None:
-                if not node.is_continuous:
-                    node.default_transistor_configuration["transistor"] = (
-                        SampleAndHoldModifier(node.step_size).apply_class(
-                            node.default_transistor_configuration["transistor"]
-                        )
-                    )
-
-                node.with_transistor(
-                    node.default_transistor_configuration["transistor"],
-                    **node.default_transistor_configuration["transistor_kwargs"],
-                )
-
-    def _log_node_order(self):
-        self.ordered_nodes_str = " -> ".join(
-            [node.state.name for node in self.ordered_nodes]
-        )
-        print(f"Resolved node order: {self.ordered_nodes_str}")
-
-    @staticmethod
-    def resolve(nodes: List[Node]) -> List[Node]:
-        # 1. Collect all full paths from each node and ensure uniqueness
-        full_path_to_node = {}
-        for node in nodes:
-            for path in node.state.paths:
-                if path in full_path_to_node:
-                    # Handle the case where different nodes produce the same full path
-                    # This is not allowed and should not be allowed
-                    other_node = full_path_to_node[path]
-                    raise ValueError(
-                        f"Duplicate full state path detected: '{path}' is produced by both "
-                        f"'{other_node.state.name}' and '{node.state.name}'. "
-                        "All full state paths must be unique."
-                    )
-                full_path_to_node[path] = node
-
-        # 2. Build the adjacency list based on full paths
-        graph = {node: [] for node in nodes}
-
-        # For each node, resolve its inputs using full paths
-        # If any input does not map to an existing path, raise an error
-        for node in nodes:
-            if not node.inputs.paths_to_states:
-                continue
-            for input_path in node.inputs.paths_to_states:
-                if input_path not in full_path_to_node:
-                    raise ValueError(
-                        f"Input path '{input_path}' required by '{node.state.name}' "
-                        "does not map to any known node's full state path."
-                    )
-                producer_node = full_path_to_node[input_path]
-                if producer_node != node:
-                    graph[producer_node].append(node)
-
-        # 3. Detect strongly connected components (SCCs) to handle cycles as blocks
-        index = 0
-        stack = []
-        on_stack = set()
-        indices = {}
-        low_link = {}
-        sccs = []
-
-        def strongconnect(v):
-            nonlocal index
-            indices[v] = index
-            low_link[v] = index
-            index += 1
-            stack.append(v)
-            on_stack.add(v)
-
-            for w in graph[v]:
-                if w not in indices:
-                    strongconnect(w)
-                    low_link[v] = min(low_link[v], low_link[w])
-                elif w in on_stack:
-                    low_link[v] = min(low_link[v], indices[w])
-
-            # If v is the root of an SCC
-            if low_link[v] == indices[v]:
-                scc = []
-                while True:
-                    w = stack.pop()
-                    on_stack.remove(w)
-                    scc.append(w)
-                    if w == v:
-                        break
-                sccs.append(scc)
-
-        for n in nodes:
-            if n not in indices:
-                strongconnect(n)
-
-        # 4. Build SCC graph (condensation)
-        node_to_scc = {}
-        for i, scc in enumerate(sccs):
-            for node in scc:
-                node_to_scc[node] = i
-
-        scc_graph = {i: set() for i in range(len(sccs))}
-        for u in nodes:
-            u_scc = node_to_scc[u]
-            for v in graph[u]:
-                v_scc = node_to_scc[v]
-                if u_scc != v_scc:
-                    scc_graph[u_scc].add(v_scc)
-
-        # 5. Topological sort on scc_graph
-        in_degree = {i: 0 for i in range(len(sccs))}
-        for u in scc_graph:
-            for v in scc_graph[u]:
-                in_degree[v] += 1
-
-        # Sort SCCs by lexicographically smallest state name for stable order
-        def scc_key(scc_index):
-            names = [node.state.name for node in sccs[scc_index]]
-            return min(names)
-
-        ready = [i for i in in_degree if in_degree[i] == 0]
-        ready.sort(key=scc_key)
-
-        scc_order = []
-        while ready:
-            current = ready.pop(0)
-            scc_order.append(current)
-            for dep in scc_graph[current]:
-                in_degree[dep] -= 1
-                if in_degree[dep] == 0:
-                    ready.append(dep)
-            ready.sort(key=scc_key)
-
-        # 6. Expand each SCC. Sort nodes within each SCC block by state name.
-        final_order = []
-        for scc_idx in scc_order:
-            scc_block = sccs[scc_idx]
-            if len(scc_block) > 1 or any(n in graph[n] for n in scc_block):
-                # Cycle or self-loop
-                scc_block.sort(key=lambda x: x.state.name)
-            else:
-                # Single node, no cycle
-                # Just stable sort by name anyway
-                scc_block.sort(key=lambda x: x.state.name)
-
-            final_order.extend(scc_block)
-
-        return final_order
-
-    def step(self):
-        """Execute a single time step for all nodes in the graph in resolved order."""
-        for node in self.ordered_nodes:
-            if node.transistor:
-                node.transistor.step()
-            else:
-                raise ValueError(f"Node {node.state.name} does not have a transistor.")
-
-    def reset(self, states_to_reset: Optional[List[str]] = None) -> None:
-        """Reset specified states or all nodes if none specified.
-
-        Args:
-            states_to_reset: List of state paths to reset. If None, resets all nodes.
-        """
-        if states_to_reset is None:
-            for node in self.nodes:
-                node.reset()
-            return
-
-        found_states = set()
-        for node in self.nodes:
-            if reset_states := node.reset(states_to_reset):
-                found_states.update(reset_states)
-
-        if missing := set(states_to_reset) - found_states:
-            raise ValueError(f"Could not find states: {missing}")
 
 
 class Clock(Node):
