+++ tests/test_node.py
@@ -1,5 +1,6 @@
 import pytest
-from regelum.environment.node.base import State, Inputs, Node, Graph, Clock
+from regelum.environment.node.base import State, Inputs, Node
+from regelum.environment.graph import Graph
 from regelum.environment.transistor import Transistor
 import numpy as np
